requirements.txt
django>=2.2.1
channels>=2.2.0
channels-redis>=2.4.0
django-cors-headers>=3.0.2
django-basicauth>=0.5.2
django-babeljs>=0.2
djangorestframework>=3.9.4
djangochannelsrestframework
